chore: reorganise wallet tests
