+++ gui/components/wallet/wallet_account_popups.py
@@ -104,7 +104,7 @@ class AccountPopup(BasePopup):
     def set_derivation_path(self, value: str, index: int, password: str):
         self._edit_derivation_path_button.hover().click()
         AuthenticatePopup().wait_until_appears().authenticate(password)
-        if value in [_.value for _ in constants.wallet.DerivationPath]:
+        if value in [_.value for _ in DerivationPath]:
             self._derivation_path_combobox_button.click()
             self._derivation_path_list_item.real_name['title'] = value
             self._derivation_path_list_item.click()
