+++ tests/settings/test_settings_sign_out_and_quit.py
