+++ tests/wallet_main_screen/test_plus_button_add_account_after_restart.py
