+++ tests/wallet_main_screen/test_saved_addresses.py
